Bug fix & enable "N/A" validation
So, when we switch scrum masters, we need to update their estimate as at the end of the day they are just a user with a flag. So, to get around this I enabled "N/A" in the validation when updating an estimate. Works flawlessly now, and should also help the 'reset estimates' side of the app.